Implemented a simple test suite
The test suite simply runs makel in turn on all Makefiles in tests/,
and compares makel's exit codes with the expected exit codes that are
written in the Makefiles themselves.

This system could be improved by comparing warning classes instead,
and using text identifiers instead of numerical codes which might be subject to change.

Revision 2 of this patch:
Make target to run the test suite is now called 'check' to avoid confusion.